refactor(!): further type-narrowing of settings
