+++ .changeset/pink-jars-admire.md
@@ -1,0 +1,5 @@
+---
+"@elgato/streamdeck": major
+---
+
+Updated the settings type used by `getSettings` and `setSettings` to use the type defined on the action.
