+++ packages/plugin/src/plugin/actions/__tests__/key.test.ts
@@ -13,7 +13,7 @@ import {
 import { connection } from "../../connection.js";
 import { Device } from "../../devices/device.js";
 import { deviceStore } from "../../devices/store.js";
-import { Action } from "../action.js";
+import { ActionBase } from "../action-base.js";
 import { KeyAction } from "../key.js";
 
 vi.mock("../../devices/store.js");
@@ -64,7 +64,7 @@ describe("KeyAction", () => {
 		const action = new KeyAction(source);
 
 		// Assert.
-		expect(action).toBeInstanceOf(Action);
+		expect(action).toBeInstanceOf(ActionBase);
 		expect(action.coordinates).not.toBeUndefined();
 		expect(action.coordinates?.column).toBe(1);
 		expect(action.coordinates?.row).toBe(2);
