+++ packages/plugin/src/plugin/actions/key.ts
@@ -2,13 +2,13 @@ import type { JsonObject } from "@elgato/utils";
 
 import type { Coordinates, SetImage, SetTitle, State, WillAppear } from "../../api/index.js";
 import { connection } from "../connection.js";
-import { Action } from "./action.js";
+import { ActionBase } from "./action-base.js";
 
 /**
  * Provides a contextualized instance of a key action.
  * @template TSettings The type of settings associated with the action.
  */
-export class KeyAction<TSettings extends JsonObject> extends Action<TSettings> {
+export class KeyAction<TSettings extends JsonObject> extends ActionBase<TSettings> {
 	/**
 	 * Private backing field for the coordinates.
 	 */
