+++ packages/plugin/src/plugin/ui.ts
@@ -70,7 +70,7 @@ class UIController {
 		listener: (ev: PropertyInspectorDidAppearEvent<T>) => void,
 	): IDisposable {
 		return connection.disposableOn("propertyInspectorDidAppear", (ev) => {
-			const action = actionStore.getActionById(ev.context);
+			const action = actionStore.getActionById<T>(ev.context);
 			if (action) {
 				listener(new ActionWithoutPayloadEvent(action, ev));
 			}
@@ -88,7 +88,7 @@ class UIController {
 		listener: (ev: PropertyInspectorDidDisappearEvent<T>) => void,
 	): IDisposable {
 		return connection.disposableOn("propertyInspectorDidDisappear", (ev) => {
-			const action = actionStore.getActionById(ev.context);
+			const action = actionStore.getActionById<T>(ev.context);
 			if (action) {
 				listener(new ActionWithoutPayloadEvent(action, ev));
 			}
@@ -106,7 +106,7 @@ class UIController {
 		listener: (ev: SendToPluginEvent<TPayload, TSettings>) => void,
 	): IDisposable {
 		return connection.disposableOn("sendToPlugin", (ev) => {
-			const action = actionStore.getActionById(ev.context);
+			const action = actionStore.getActionById<TSettings>(ev.context);
 			if (action) {
 				listener(
 					new SendToPluginEvent<TPayload, TSettings>(action, ev as DidReceivePropertyInspectorMessage<TPayload>),
